Привел переменые balance, transfer к типу int64

+++ main.go
@@ -3,6 +3,6 @@ package main
 func main() {
 	var balance int32 = 15_000_000_00
 	var transfer int32 = 10_000_000_00
-	total := balance + transfer
+	total := int64(balance) + int64(transfer)
 	println(total)
 }
